+++ Chilicki.Commline/Chilicki.Commline.UserInterface/HtmlExtensions/HtmlStopTypeExtensions.cs
@@ -13,7 +13,7 @@ namespace Chilicki.Commline.UserInterface.HtmlExtensions
         public static SelectList ToSelectList(this StopType selectedValueOutput)
         {
             var values = from StopType e in Enum.GetValues(typeof(StopType))
-                         select new { Id = e, Name = e.GetDescription() };
+                         select new { Id = (int)e, Name = e.GetDescription() };
             return new SelectList(values, "Id", "Name", selectedValueOutput);
         }
